Improve stability of logarithms in the full-matrix QRE implementation
For some reason taking the logarithms after diagonalizing and then
rotating back was giving weird numerical issues. Not only limited to
getting anomalous negative values for the QRE sometimes, but actually
finding genuinely false minima (lower than the true global minimum...).

Matlab's native logm function appears to solve the issue, as tested on
data for the single-site RDM in the attractive Hubbard model.

We might want to investigate in more depth the issue with the naive
implementation, so I leave it commented instead of deleting it, for now.