It appears that for actual non-classical states we need Haar guesses
Starting with the identity makes us converge much faster, but to a
greatly enhanced value of the discord, on data for the single-site
RDM in the attractive Hubbard model. This makes sense, as the closest
pseudo-classical state is quite different from the input RDM, as it
has the same diagonal, but removes the pairing amplitude from the
holon-doublon off-diagonal terms.